moved scene info class
