+++ Plugin/Trackers/SceneIndexNameTracker/SceneIndexNameTracker.cs
@@ -67,16 +67,4 @@ public class SceneIndexNameTracker : IPluginInitialLoad, ISceneIndexName, IOnUpd
     {
         return _sceneInfos.Find(x => x.SceneIndex == sceneIndex)?.SceneName;
     }
-
-    private class SceneInfo
-    {
-        public int SceneIndex { get; }
-        public string SceneName { get; }
-
-        public SceneInfo(int sceneIndex, string sceneName)
-        {
-            SceneIndex = sceneIndex;
-            SceneName = sceneName;
-        }
-    }
 }
